+++ core/src/logging.rs
@@ -6,12 +6,44 @@
 //! Everything related to logging.
 use std::fs::File;
 use std::path::{Path, PathBuf};
+use std::sync::{LazyLock, Mutex};
 
 use anyhow::{Context, Result};
-use log::LevelFilter;
+use hashbrown::HashMap;
+use log::{warn, LevelFilter};
 use simplelog::{
     ColorChoice, CombinedLogger, Config, SharedLogger, TermLogger, TerminalMode, WriteLogger,
 };
+
+static SENT_WARNINGS: LazyLock<Mutex<HashMap<WarningType, usize>>> =
+    LazyLock::new(|| Mutex::new(HashMap::new()));
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+/// Enum representing the various type of warning messages that can be sent.
+pub(crate) enum WarningType {
+    /// Warning for the AlphaBetaGamma schedule-utility type being deprecated.
+    DeprecatedAlphaBetaGamma,
+}
+
+/// Sends a warning message if it was sent less than `n` times before.
+pub(crate) fn send_warning_at_most_n_times(warn_type: WarningType, message: &str, n: usize) {
+    let mut sent_warnings = LazyLock::force(&SENT_WARNINGS).lock().unwrap();
+    match sent_warnings.get(&warn_type) {
+        Some(count) if *count >= n => {
+            // The warning message was already sent at least `n` times.
+        }
+        _ => {
+            // Send the warning message.
+            warn!("{}", message);
+            *sent_warnings.entry(warn_type).or_insert(0) += 1;
+        }
+    }
+}
+
+/// Sends a warning message if it was never sent before.
+pub(crate) fn send_warning_at_most_once(warn_type: WarningType, message: &str) {
+    send_warning_at_most_n_times(warn_type, message, 1)
+}
 
 /// Initializes logging to a file and terminal.
 pub fn initialize_logging<W: std::io::Write + Send + 'static>(