+++ core/src/schedule_utility/mod.rs
@@ -4,11 +4,14 @@
 // https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode
 
 //! Everything related to schedule utility.
-use alpha_beta_gamma::AlphaBetaGammaModel;
+use alpha_beta_gamma::LinearPenaltiesModel;
 use anyhow::{anyhow, bail, Context, Result};
 use num_traits::ConstZero;
 
-use crate::units::{NonNegativeSeconds, Utility};
+use crate::{
+    logging::{send_warning_at_most_once, WarningType},
+    units::{NonNegativeSeconds, Utility},
+};
 
 pub mod alpha_beta_gamma;
 
@@ -24,7 +27,7 @@ pub enum ScheduleUtility {
     ///
     /// There is a penalty beta for leaving / arriving early and a penalty gamma for leaving /
     /// arriving late.
-    AlphaBetaGamma(AlphaBetaGammaModel),
+    LinearPenalties(LinearPenaltiesModel),
 }
 
 impl Default for ScheduleUtility {
@@ -42,14 +45,26 @@ impl ScheduleUtility {
         delta: Option<f64>,
     ) -> Result<Self> {
         match utility_type {
+            Some("Linear") => {
+                let model = LinearPenaltiesModel::from_values(tstar, beta, gamma, delta)
+                    .with_context(|| {
+                        anyhow!("Failed to create schedule utility with `type`=`\"Linear\"`")
+                    })?;
+                Ok(ScheduleUtility::LinearPenalties(model))
+            }
             Some("AlphaBetaGamma") => {
-                let model = AlphaBetaGammaModel::from_values(tstar, beta, gamma, delta)
+                let model = LinearPenaltiesModel::from_values(tstar, beta, gamma, delta)
                     .with_context(|| {
                         anyhow!(
                             "Failed to create schedule utility with `type`=`\"AlphaBetaGamma\"`"
                         )
                     })?;
-                Ok(ScheduleUtility::AlphaBetaGamma(model))
+                send_warning_at_most_once(
+                    WarningType::DeprecatedAlphaBetaGamma,
+                    "Schedule utility type \"AlphaBetaGamma\" is deprecated. \
+                    Use \"Linear\" instead.",
+                );
+                Ok(ScheduleUtility::LinearPenalties(model))
             }
             None => Ok(ScheduleUtility::None),
             Some(s) => bail!("Unknown type: {s}"),
@@ -61,7 +76,7 @@ impl ScheduleUtility {
     /// The breakpoints are ordered by increasing departure time.
     pub fn iter_breakpoints(&self) -> Box<dyn Iterator<Item = NonNegativeSeconds>> {
         match self {
-            Self::AlphaBetaGamma(model) => model.iter_breakpoints(),
+            Self::LinearPenalties(model) => model.iter_breakpoints(),
             _ => Box::new(std::iter::empty()),
         }
     }
@@ -71,7 +86,7 @@ impl ScheduleUtility {
     pub fn get_utility(&self, time: NonNegativeSeconds) -> Utility {
         match self {
             Self::None => Utility::ZERO,
-            Self::AlphaBetaGamma(model) => model.get_utility(time),
+            Self::LinearPenalties(model) => model.get_utility(time),
         }
     }
 }